chore: factor out mcpservers
