+++ src/services/claudeService.ts
@@ -1,7 +1,3 @@
-import fs from "node:fs";
-import path from "node:path";
-import { fileURLToPath } from "node:url";
-
 import {
   query,
   type SDKMessage,
@@ -10,11 +6,10 @@ import {
   type SDKResultMessage,
   type SDKSystemMessage,
   type PermissionMode,
-  type Options,
-  McpStdioServerConfig,
 } from "@anthropic-ai/claude-code";
 
 import { readSelectedModelFromConfig } from "./config.js";
+import { createMCPServer } from "./mcp.js";
 import { SYSTEM_PROMPT } from "./prompts.js";
 
 export interface ClaudeOptions {
@@ -120,52 +115,20 @@ class MessageRenderer {
 }
 
 export class ClaudeService {
-  private static getMcpServerPath(): string {
-    const __filename = fileURLToPath(import.meta.url);
-    const __dirname = path.dirname(__filename);
-
-    // Try compiled version first (dist/server.js)
-    const compiledServerPath = path.resolve(__dirname, "..", "server.js");
-    if (!fs.existsSync(compiledServerPath)) {
-      throw new Error("Compiled MCP server not found. Run `npm run build` first.");
-    }
-
-    return compiledServerPath;
-  }
-
-  private static createQueryOptions(params: ClaudeOptions): Partial<Options> {
-    const selectedModel = readSelectedModelFromConfig();
-
-    const serverPath = this.getMcpServerPath();
-    const mcpServers = {
-      "finance-agent": {
-        type: "stdio",
-        command: "node",
-        args: [serverPath],
-        env: {
-          ...process.env,
-          PWD: process.cwd(),
-        },
-      } as McpStdioServerConfig,
-    };
-
-    const baseOptions: Partial<Options> = {
-      model: selectedModel,
-      abortController: params.abortController,
-      customSystemPrompt: SYSTEM_PROMPT,
-      // TODO: add proper permission handling
-      permissionMode: "bypassPermissions" as PermissionMode,
-      allowedTools: Object.keys(mcpServers).map((name) => `mcp__${name}`),
-      mcpServers,
-    };
-
-    return baseOptions;
-  }
-
   static async executePrompt(prompt: string, options: ClaudeOptions = {}): Promise<ClaudeResponse> {
     try {
       const abortController = options.abortController || new AbortController();
-      const queryOptions = this.createQueryOptions({ abortController });
+      const mcpServers = createMCPServer();
+
+      const queryOptions = {
+        model: readSelectedModelFromConfig(),
+        abortController,
+        customSystemPrompt: SYSTEM_PROMPT,
+        // TODO: add proper permission handling
+        permissionMode: "bypassPermissions" as PermissionMode,
+        allowedTools: Object.keys(mcpServers).map((name) => `mcp__${name}`),
+        mcpServers,
+      };
 
       for await (const message of query({
         prompt,
